# server/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DailyLogSummary": {
        "description": "Response fragment summarizing one day's totals and entry count.",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "entry_count": {
            "title": "Entry Count",
            "type": "integer"
          },
          "excluded": {
            "default": false,
            "title": "Excluded",
            "type": "boolean"
          },
          "total_calories": {
            "title": "Total Calories",
            "type": "integer"
          },
          "total_carbs_g": {
            "title": "Total Carbs G",
            "type": "number"
          },
          "total_fat_g": {
            "title": "Total Fat G",
            "type": "number"
          },
          "total_protein_g": {
            "title": "Total Protein G",
            "type": "number"
          }
        },
        "required": [
          "date",
          "total_calories",
          "total_protein_g",
          "total_carbs_g",
          "total_fat_g",
          "entry_count"
        ],
        "title": "DailyLogSummary",
        "type": "object"
      },
      "DailySummaryResponse": {
        "description": "Response body for ``GET /summary?date=...`` \u2014 full daily macro picture.\n\nThe REST ``/summary`` endpoint always populates ``target``/``remaining`` (it\n404s when no profile exists). They are nullable only to support the MCP\n``get_day`` tool's no-target mode, which returns consumed totals with\n``target``/``remaining`` set to ``None``.",
        "properties": {
          "consumed": {
            "$ref": "#/components/schemas/MacroTotals"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FoodEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "excluded": {
            "default": false,
            "title": "Excluded",
            "type": "boolean"
          },
          "remaining": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MacroTotals"
              },
              {
                "type": "null"
              }
            ]
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MacroTargets"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "date",
          "target",
          "consumed",
          "remaining",
          "entries"
        ],
        "title": "DailySummaryResponse",
        "type": "object"
      },
      "DayExclusionRequest": {
        "description": "Request body for ``PUT /logs/{date}/excluded`` \u2014 the new flag value.",
        "properties": {
          "excluded": {
            "title": "Excluded",
            "type": "boolean"
          }
        },
        "required": [
          "excluded"
        ],
        "title": "DayExclusionRequest",
        "type": "object"
      },
...
      "FoodEntryResponse": {
        "description": "Response body representing one persisted food entry row.",
        "properties": {
          "calories": {
            "title": "Calories",
            "type": "integer"
          },
          "carbs_g": {
            "title": "Carbs G",
            "type": "number"
          },
          "confirmed": {
            "default": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "consumed_at": {
            "format": "date-time",
            "title": "Consumed At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "custom_food_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Food Id"
          },
          "daily_log_id": {
            "format": "uuid",
            "title": "Daily Log Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "entry_group_id": {
            "format": "uuid",
            "title": "Entry Group Id",
            "type": "string"
          },
          "fat_g": {
            "title": "Fat G",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "meal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Id"
          },
          "meal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Name"
          },
          "normalized_quantity_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Quantity Unit"
          },
          "normalized_quantity_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Quantity Value"
          },
          "protein_g": {
            "title": "Protein G",
            "type": "number"
          },
          "quantity_text": {
            "title": "Quantity Text",
            "type": "string"
          },
          "usda_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usda Description"
          },
          "usda_fdc_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usda Fdc Id"
          },
          "user_key": {
            "title": "User Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "daily_log_id",
          "user_key",
          "entry_group_id",
          "display_name",
          "quantity_text",
          "normalized_quantity_value",
          "normalized_quantity_unit",
          "calories",
          "protein_g",
          "carbs_g",
          "fat_g",
          "consumed_at",
          "created_at"
        ],
        "title": "FoodEntryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LogMealRequest": {
        "description": "Request body for ``POST /meals/{id}/log``.\n\nA naive ``consumed_at`` is treated as wall-clock time in the configured\ntimezone (``TZ``); ``None`` defers to the server's current time.",
        "properties": {
          "consumed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumed At"
          }
        },
        "title": "LogMealRequest",
        "type": "object"
      },
      "LogMealResponse": {
        "description": "Response body for ``POST /meals/{id}/log``: created entries plus updated daily totals.",
        "properties": {
          "daily_totals": {
            "$ref": "#/components/schemas/MacroTotals"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FoodEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries",
          "daily_totals"
        ],
        "title": "LogMealResponse",
        "type": "object"
      },
      "LogsListResponse": {
        "description": "Response body for ``GET /logs`` \u2014 wraps a series of daily summaries.",
        "properties": {
          "logs": {
            "items": {
              "$ref": "#/components/schemas/DailyLogSummary"
            },
            "title": "Logs",
            "type": "array"
          }
        },
        "required": [
          "logs"
        ],
        "title": "LogsListResponse",
        "type": "object"
      },
      "MacroTargets": {
        "description": "User-configured daily macro targets plus an optional target weight.\n\nRequest/response body for the targets endpoints. ``target_weight_lb``\nis stored as ``numeric(6,2)`` in the database (max ``9999.99``) and\nmust round to at least ``0.01`` when provided.",
        "properties": {
          "calories": {
            "exclusiveMinimum": 0.0,
            "title": "Calories",
            "type": "integer"
          },
          "carbs_g": {
            "minimum": 0.0,
            "title": "Carbs G",
            "type": "number"
          },
          "fat_g": {
            "minimum": 0.0,
            "title": "Fat G",
            "type": "number"
          },
          "protein_g": {
            "minimum": 0.0,
            "title": "Protein G",
            "type": "number"
          },
          "target_weight_lb": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 9999.99,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Weight Lb"
          }
        },
        "required": [
          "calories",
          "protein_g",
          "carbs_g",
          "fat_g"
        ],
        "title": "MacroTargets",
        "type": "object"
      },
      "MacroTotals": {
        "description": "Aggregated macro totals (calories + protein/carbs/fat in grams).\n\nUsed as a response-body fragment wherever the API returns \"how much\nhas been consumed / remains / is targeted\" \u2014 e.g. daily summaries\nand entry-creation responses.",
        "properties": {
          "calories": {
            "title": "Calories",
            "type": "integer"
          },
          "carbs_g": {
            "title": "Carbs G",
            "type": "number"
          },
          "fat_g": {
            "title": "Fat G",
            "type": "number"
          },
          "protein_g": {
            "title": "Protein G",
            "type": "number"
          }
        },
        "required": [
          "calories",
          "protein_g",
          "carbs_g",
          "fat_g"
        ],
        "title": "MacroTotals",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Diet Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/logs": {
      "get": {
        "description": "List historical daily logs for the authenticated user across an inclusive date range.\n\n**Inputs:**\n- request (Request): Active request providing ``user_key``.\n- from_date (date): Inclusive start date (query alias ``from``).\n- to_date (date): Inclusive end date (query alias ``to``).\n- session (AsyncSession): DB session dependency.\n\n**Outputs:**\n- LogsListResponse: Daily aggregate totals and entry counts ordered by date descending.\n\n**Exceptions:**\n- HTTPException(400): Raised when ``from_date`` is after ``to_date`` or the\n  span exceeds ``MAX_RANGE_DAYS``.\n- RuntimeError: Raised when the database pool is not initialized.\n- sqlalchemy.exc.SQLAlchemyError: Raised when SQL execution fails.",
        "operationId": "list_logs_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Logs"
      }
    },
    "/logs/{log_date}/excluded": {
      "put": {
        "description": "Toggle the manual \"ignore this day from stats\" flag for one calendar date.\n\nUpserts the day's ``daily_logs`` row (so a never-logged day can still be\nexcluded), then returns the day's refreshed summary carrying the new\n``excluded`` value. Setting the flag never alters the day's own entries or\nconsumed totals \u2014 only whether aggregates skip it.\n\n**Inputs:**\n- request (Request): Active request providing ``user_key``.\n- log_date (date): Calendar date to toggle (path parameter).\n- body (DayExclusionRequest): New flag value (``{\"excluded\": bool}``).\n- session (AsyncSession): DB session dependency.\n\n**Outputs:**\n- DailySummaryResponse: The day's summary after the update, with\n  ``target``/``remaining`` populated and ``excluded`` reflecting the new\n  value. Mirrors the ``GET /summary`` contract (always-populated target).\n\n**Exceptions:**\n- HTTPException(404): Raised when no target profile exists for the user\n  (same contract as ``GET /summary``).\n- sqlalchemy.exc.SQLAlchemyError: Raised when the write or read fails.",
        "operationId": "set_day_excluded_logs__log_date__excluded_put",
        "parameters": [
          {
            "in": "path",
            "name": "log_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Log Date",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayExclusionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Day Excluded"
      }
    },
...
    "/meals/{meal_id}/log": {
      "post": {
        "description": "Expand every meal item into a food entry sharing one ``entry_group_id``.\n\nA naive ``consumed_at`` is interpreted in the configured timezone; ``None``\ndefers to the server's current wall-clock time.\n\n**Inputs:**\n- request (Request): Active request providing ``user_key``.\n- meal_id (UUID): Meal to log.\n- body (LogMealRequest | None): Optional payload carrying ``consumed_at``.\n- session (AsyncSession): DB session dependency.\n\n**Outputs:**\n- LogMealResponse: The created entries plus updated daily macro totals.",
        "operationId": "log_meal_endpoint_meals__meal_id__log_post",
        "parameters": [
          {
            "in": "path",
            "name": "meal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Meal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LogMealRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogMealResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Meal Endpoint"
      }
    },
...
  }
}